Move `plot_compare` out of `helpers` into a module of its own
`plot_compare` and the two dozen private helpers it is built from had grown to
be almost all of `helpers`, which is otherwise a couple of small conveniences.

Move it out as it is. `helpers` keeps `plot_arrows` and `view_vectors`, and the
imports it no longer needs are dropped. The deprecated `plot_compare_four` is
annotated on the way, since the new module has no exemption from annotating
what it defines, and `plot_compare` is documented as `pyvista.plot_compare`
either way, so nothing moves as far as the docs are concerned.
