+++ pyvista/plotting/__init__.py
@@ -56,8 +56,8 @@ from .errors import InvalidCameraError as InvalidCameraError
 from .errors import RenderWindowUnavailable as RenderWindowUnavailable
 from .follower import Follower as Follower
 from .helpers import plot_arrows as plot_arrows
-from .helpers import plot_compare as plot_compare
-from .helpers import plot_compare_four as plot_compare_four
+from .plot_compare import plot_compare as plot_compare
+from .plot_compare import plot_compare_four as plot_compare_four
 from .interactor_style_registry import register_interactor_style as register_interactor_style
 from .lights import Light as Light
 from .lookup_table import LookupTable as LookupTable
